correct ddock size

+++ src/Net.cpp
@@ -70,7 +70,7 @@ TrainResult Net::train(const Matrix& mSamples,const Matrix& mTruth,const TrainOp
                 //compute one sample error
 
                 //forward pass with save
-                int iIndexSample=mShuffle(iSample);
+                int iIndexSample=(int)mShuffle(iSample);
                 const Matrix& mSample=mSamples.row(iIndexSample);
                 Matrix mOut, mTemp=mSample;
                 for(unsigned int i=0;i<_layers.size();i++)
